pyproject.toml
[tool.poetry]
name = "SteamMarketBot"
version = "0.1.0"
description = ""
authors = ["_Counter021_"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^3.141.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
